% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group.preferences.R
\name{group}
\alias{group}
\alias{as.grouped_preferences}
\alias{as.grouped_preferences.paircomp}
\alias{[.grouped_preferences}
\alias{format.grouped_preferences}
\title{Group Preferences}
\usage{
group(x, index, ...)

as.grouped_preferences(x, ...)

\method{as.grouped_preferences}{paircomp}(x, ...)

\method{[}{grouped_preferences}(x, i, j, ..., drop = TRUE, as.grouped_preferences = TRUE)

\method{format}{grouped_preferences}(x, max = 2L, width = 20L, ...)
}
\arguments{
\item{x}{A \code{\link[=preferences]{"preferences"}} object for \code{group()}; an object
that can be coerced to a \code{"grouped_preferences"} object for
\code{as.grouped_preferences()}, otherwise a \code{"grouped_preferences"} object.}

\item{index}{A numeric vector of length equal to the number of preferences
specifying the subject for each ranking.}

\item{...}{Additional arguments passed on to \verb{\link{as.preferences}}
by \code{grouped_preferences} or \code{as.grouped_preferences}; unused by
\code{format}.}

\item{i}{Indices specifying groups to extract, may be any data type accepted
by \verb{\link{[}}.}

\item{j}{Indices specifying items to extract, as for \verb{\link{[}}.}

\item{drop}{If \code{TRUE}, return single row/column matrices as a vector.}

\item{as.grouped_preferences}{If \code{TRUE} return a \code{"grouped_preferences"}
object, otherwise return a matrix/vector.}

\item{max}{The maximum number of preferences to format per subject.}

\item{width}{The maximum width in number of characters to format each
ranking.}
}
\value{
An object of class \code{"grouped_preferences"}, which is a vector of
of group IDs with the following attributes:
\item{preferences}{ The \code{"preferences"} object.}
\item{index}{ An index match each ranking to each group ID.}
\item{R}{ A matrix with items ordered from last to first place, for each
ranking.}
\item{S}{ The preferences matrix with the ranks replaced by the size of the
chosen set for free choices and zero for forced choices.}
\item{id}{ A list with elements of the adjacency matrix that are incremented
by each ranking.}
}
\description{
Create an object of class \code{"grouped_preferences"} which associates a
group index with an object of class \code{"preferences"}. This allows the
preferences to be linked to covariates with group-specific values.
}
\examples{

# ungrouped preferences (5 preference sets, 4 items)
R <- as.preferences(matrix(c(1, 2, 0, 0,
                             0, 2, 1, 0,
                             0, 0, 1, 2,
                             2, 1, 0, 0,
                             0, 1, 2, 3), ncol = 4, byrow = TRUE))
length(R)
R

# group preferences (first three in group 1, next two in group 2)
G <- group(R, c(1, 1, 1, 2, 2))
length(G)

## by default up to 2 preference sets are shown per group, "..." indicates if
## there are further preferences
G
print(G, max = 1)

## select preferences from group 1
G[1,]

## exclude item 3 from ranking
G[, -3]

## preferences from group 2, excluding item 3
## - note group 2 becomes the first group
G[2, -3]

## index underlying preferences without creating new grouped_preferences
## object
G[2, -3, as.grouped_preferences = FALSE]
}
