% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{items_at_rank}
\alias{items_at_rank}
\title{Get the name of the item(s) assigned a specific rank, e.g., first.}
\usage{
items_at_rank(x, rank)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}

\item{rank}{A single integer, the rank which you would like to inspect.}
}
\value{
A list containing the name(s) of the item(s) ranked \code{rank} in each of
the preferences in \code{x}.
}
\description{
Get the name of the item(s) assigned a specific rank, e.g., first.
}
