% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vctrs-prefio.R
\name{new_prefio}
\alias{new_prefio}
\title{Create a new vector of preferences}
\usage{
new_prefio(x, levels)
}
\arguments{
\item{x}{The list of integers, containing the indexed preferences in order}

\item{levels}{The labels of the preferences}
}
\description{
Create a new vector of preferences
}
