% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_blank}
\alias{pref_blank}
\title{Check if a preference is blank.}
\usage{
pref_blank(x)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}
}
\value{
A logical vector indicating which preferences are blank, i.e., \verb{[]}.
}
\description{
Check if a preference is blank.
}
