% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_complete}
\alias{pref_complete}
\title{Complete preferences by adding unselected items as last place occurrances.}
\usage{
pref_complete(x)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}
}
\value{
A new vector of preferences, with each selection starting with the
corresponding selections made in \code{x}, but with all unranked items placed
last.
}
\description{
Complete preferences by adding unselected items as last place occurrances.
}
\examples{
# Complete partial rankings by adding unranked items last
pref_complete(preferences(c("a > b", "c > a", "b")))
}
