% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_summaries.R
\name{pref_irv}
\alias{pref_irv}
\title{Compute the instant-runoff voting winner for a set of preferences.}
\usage{
pref_irv(x, preferences_col = NULL, frequency_col = NULL)
}
\arguments{
\item{x}{A vector of preferences, or a tibble with a column of preferences.}

\item{preferences_col}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> When \code{x} is a
\code{tibble}, the column containing the preferences.}

\item{frequency_col}{<\code{\link[dplyr:dplyr_tidy_select]{tidy-select}}> When \code{x} is a
\code{tibble}, the column containing the frequency of the preferences. If not
provided, each row is considered to be observed a single time.}
}
\value{
A summary of the results of the IRV contest:
...
}
\description{
Compute the instant-runoff voting winner for a set of preferences.
}
\examples{
# Simple contest with five votes between three candidates a, b, and c.
pref_irv(preferences(c("a > b > c", "a > b > c", "b > c > a", "b > a > c", "b > a > c")))

# Weighted covariance by frequency
df <- tibble::tibble(
  prefs = preferences(c("a > b > c", "b > c > a", "c > a > b")),
  freq = c(3, 2, 1)
)
pref_irv(df, preferences_col = prefs, frequency_col = freq)
}
