% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_items_at_rank}
\alias{pref_items_at_rank}
\title{Get the name of the item(s) assigned a specific rank, e.g., first.}
\usage{
pref_items_at_rank(x, rank)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}

\item{rank}{A single integer, the rank which you would like to inspect.}
}
\value{
A list containing the name(s) of the item(s) ranked \code{rank} in each of
the preferences in \code{x}.
}
\description{
Get the name of the item(s) assigned a specific rank, e.g., first.
}
\examples{
# Get items ranked first
pref_items_at_rank(preferences(c("a > b > c", "b = c > a")), rank = 1)
# Get items ranked second
pref_items_at_rank(preferences(c("a > b > c", "b = c > a")), rank = 2)
}
