% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_keep}
\alias{pref_keep}
\title{Keep only specified items from preferences.}
\usage{
pref_keep(x, items)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}

\item{items}{The names of the items which should be kept for preferences in \code{x}.}
}
\value{
A new vector of preferences, but only containing \code{items} from each selection.
}
\description{
Keep only specified items from preferences.
}
\examples{
# Keep only 'a' and 'c'
pref_keep(preferences(c("a > b > c", "b > c > a")), c("a", "c"))
}
