% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_length}
\alias{pref_length}
\title{Check the length (number of rankings) of a preference.}
\usage{
pref_length(x)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}
}
\value{
The number of items listed on each of the preferences.
}
\description{
Check the length (number of rankings) of a preference.
}
\examples{
pref_length(preferences(c("a > b > c", "", "b > c")))
}
