% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_project}
\alias{pref_project}
\title{Remove all but specified items from preferences.}
\usage{
pref_project(x, items)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}

\item{items}{The names of the items which should be kept from the preferences in \code{x}.}
}
\value{
A new vector of preferences, but with any item not in \code{items}
removed from each selection.
}
\description{
Remove all but specified items from preferences.
}
