% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_rank_of_item}
\alias{pref_rank_of_item}
\title{Get the rank assigned to a specific item in a set of preferences.}
\usage{
pref_rank_of_item(x, item_name)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}

\item{item_name}{The name of the item to extract the rank for.}
}
\value{
The rank of \code{item_name} for each of the preferences in \code{x}.
}
\description{
Get the rank assigned to a specific item in a set of preferences.
}
\examples{
pref_rank_of_item(preferences(c("a > b > c", "b > c = a", "")), "a")
}
