% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_reverse}
\alias{pref_reverse}
\title{Reverse preference rankings}
\usage{
pref_reverse(x, ...)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}

\item{...}{Not used.}
}
\value{
A vector of preferences with rankings reversed (first becomes last, etc.)
}
\description{
Reverse preference rankings
}
\examples{
reverse_preferences(preferences(c("a > b > c", "b > c > a")))
}
