% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_trunc}
\alias{pref_trunc}
\title{Truncate preferences to a maximum number of ranks.}
\usage{
pref_trunc(x, n_ranks = 1L, top_ranks = TRUE)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}

\item{n_ranks}{The number of rankings to include in the output.}

\item{top_ranks}{If \code{TRUE}, output the top \code{n_ranks}, otherwise the bottom
\code{n_ranks}.}
}
\value{
A vector of preferences with each selection equal to \code{x} but omitting
the bottom (if \code{top_ranks} is \code{TRUE}, otherwise top) selections so that the
new vector has at most \code{n_ranks} selections.
}
\description{
Truncate preferences to a maximum number of ranks.
}
\examples{
# Keep only the top 2 ranked items
pref_trunc(preferences(c("a > b > c > d", "b > c > a")), n_ranks = 2)
# Keep only the bottom 2 ranked items
pref_trunc(preferences(c("a > b > c > d", "b > c > a")), n_ranks = 2, top_ranks = FALSE)
}
