% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefio-package.R
\docType{package}
\name{prefio-package}
\alias{prefio}
\alias{prefio-package}
\title{prefio: Structures for Preference Data}
\description{
Convenient structures for creating, sourcing and manipulating preference data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fleverest/prefio/}
  \item Report bugs at \url{https://github.com/fleverest/prefio/issues/}
}

}
\author{
\strong{Maintainer}: Floyd Everest \email{me@floydeverest.com} (\href{https://orcid.org/0000-0002-2726-6736}{ORCID}) [contributor]

Authors:
\itemize{
  \item Heather Turner \email{ht@heatherturner.net} (\href{https://orcid.org/0000-0002-1256-3375}{ORCID})
}

}
\keyword{internal}
