% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preferences.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{vec_ptype_abbr}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{vctrs}{\code{\link[vctrs:vec_ptype_full]{vec_ptype_abbr}}}
}}

