% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{reverse_preferences}
\alias{reverse_preferences}
\title{Reverse preference rankings}
\usage{
reverse_preferences(x, ...)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}
}
\value{
A vector of preferences with rankings reversed (first becomes last, etc.)
}
\description{
Reverse preference rankings
}
\examples{
reverse_preferences(c("a > b > c", "b > c > a"))
reverse_preferences(c("a > b > c", "b > c > a"), format = "long")
}
