% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{rm_items}
\alias{rm_items}
\title{Remove specified items from preferences.}
\usage{
rm_items(x, items)
}
\arguments{
\item{x}{A vector of \code{\link{preferences}}.}

\item{items}{The names of the items which should be removed from the preferences in \code{x}.}
}
\value{
A new vector of preferences, but with \code{items} removed from each selection.
}
\description{
Remove specified items from preferences.
}
