% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pref_tools.R
\name{pref_cov}
\alias{pref_cov}
\title{Covariance matrix for preferences, calculated using the rankings matrix.}
\usage{
pref_cov(x, ...)
}
\arguments{
\item{x}{A vector of preferences.}

\item{...}{Extra arguments to be passed to \code{stats::cov.wt}.}
}
\value{
A covariance matrix containing covariances for the ranks assigned to
item pairs.
}
\description{
Covariance matrix for preferences, calculated using the rankings matrix.
}
