% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prefio-package.R
\docType{package}
\name{prefio-package}
\alias{prefio}
\alias{prefio-package}
\title{prefio: Structures for Preference Data}
\description{
Convenient structures for creating, sourcing, reading, writing and manipulating ordinal preference data. Methods for writing to/from PrefLib formats. See Nicholas Mattei and Toby Walsh "PrefLib: A Library of Preference Data" (2013) \doi{10.1007/978-3-642-41575-3_20}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fleverest/prefio/}
  \item \url{https://fleverest.github.io/prefio/}
  \item Report bugs at \url{https://github.com/fleverest/prefio/issues/}
}

}
\author{
\strong{Maintainer}: Floyd Everest \email{me@floydeverest.com} (\href{https://orcid.org/0000-0002-2726-6736}{ORCID})

Authors:
\itemize{
  \item Heather Turner \email{ht@heatherturner.net} (\href{https://orcid.org/0000-0002-1256-3375}{ORCID})
  \item Damjan Vukcevic \email{damjan@vukcevic.net} (\href{https://orcid.org/0000-0001-7780-9586}{ORCID})
}

}
\keyword{internal}
